Restructure otb_editor
