+++ otb_editor/components/LatLngEditor.tsx
