+++ otb_editor/pages/project/ProjectAssetsEditor.tsx
