+++ otb_editor/pages/project/TourEditorMap.tsx
