+++ otb_editor/pages/project/WaypointEditorPanel.module.css
